[DSCP-67] Add Obj type and refactor code
