+++ src/Disciplina/DB.hs
@@ -3,6 +3,5 @@ module Disciplina.DB (module M) where
 
 import Disciplina.DB.Class as M
 import Disciplina.DB.Real as M
-import Disciplina.DB.DSL.Interpret as M
-import Disciplina.DB.DSL.Types as M
-import Disciplina.DB.DSL.SimpleTxDB as M
+import Disciplina.DB.DSL.Class as M
+import Disciplina.DB.DSL.Interpret.SimpleTxDB as M
