+++ src/Disciplina/DB/DSL/Interpret/SimpleTxDB.hs
@@ -1,31 +1,36 @@
-module Disciplina.DB.DSL.SimpleTxDB
+module Disciplina.DB.DSL.Interpret.SimpleTxDB
        (runSimpleTxDBQuery
        ) where
 
 import Universum
 
 import Codec.Serialise (Serialise(..))
+import Control.Lens (traversed, filtered, makePrisms)
 
-import Disciplina.DB.DSL.Types (QueryTx(..), QueryTxs(..), WHERE(..)
-                               ,TxIdEq(..), TxsFilterExpr(..))
+import Disciplina.DB.DSL.Class (QueryTx(..), QueryTxs(..), ObjHashEq(..), Obj
+                               ,QueryObj(..), MonadSearchTxObj(..), RunQuery(..)
+                               ,WHERE(..), TxIdEq(..), TxsFilterExpr(..))
 import Disciplina.Educator.Txs (PrivateTxId, PrivateTx(..), EducatorTxMsg(..)
                                ,StudentTxMsg(..), PrivateTxPayload(..))
 import Disciplina.Crypto (hash)
 import Disciplina.Core (Address(..), CourseId(..), Grade(..)
                        , hasPathFromTo, activityTypeGraphIndexed)
-import Disciplina.DB.DSL.Interpret (MonadSearchTxObj(..), RunQuery(..))
 import Data.List (union, intersect)
 
+data SimpleObj = SSTx { _sstorePTx :: PrivateTx}
+               | SSObj { _sstoreObj :: Obj }
+
+makePrisms ''SimpleObj
 
 -- | Simple transaction database
 newtype SimpleTxDB a = SimpleTxDB
-   { getSimpleTxDB :: Reader [PrivateTx] a
-   } deriving (Functor, Applicative, Monad, MonadReader [PrivateTx])
+   { getSimpleTxDB :: Reader [SimpleObj] a
+   } deriving (Functor, Applicative, Monad, MonadReader [SimpleObj])
 
 instance MonadSearchTxObj SimpleTxDB where
   runTxQuery = evalSimpleTxQuery
   runTxsQuery = evalSimpleTxsQuery
-  runObjQuery = undefined -- TODO, implement this one, but how are Object defined?
+  runObjQuery = evalSimpleObjQuery
 
 instance Serialise Address where
   encode (Address h) = encode h
@@ -42,23 +47,39 @@ instance Serialise PrivateTx
 
 -- | Evaluator for query: find Tx in db with hash == h
 evalSimpleTxQuery :: QueryTx -> SimpleTxDB (Maybe PrivateTx)
-evalSimpleTxQuery (SELECTTx _ (TxIdEq (h :: PrivateTxId))) =
-  safeHead.filter ((==h).hash) <$> ask
+evalSimpleTxQuery (SELECTTx _ (TxIdEq (h :: PrivateTxId))) = do
+  db <- ask
+  return $ db ^?  traversed . _SSTx . filtered (((h==).hash))
+
+-- | Evaluator for query: find Tx in db with obj hash == h
+evalSimpleObjQuery :: QueryObj -> SimpleTxDB (Maybe Obj)
+evalSimpleObjQuery (SELECTObj _ (ObjHashEq h)) = do
+  db <- ask
+  return $ db ^?  traversed . _SSObj . filtered (((h==).hash))
 
 -- | Evaluator for query: find Txs in db with SubjectId == a
 evalSimpleTxsQuery :: QueryTxs -> SimpleTxDB [PrivateTx]
-evalSimpleTxsQuery (SELECTTxs _ (TxSubjectIdEq sId)) =
-  filter ((==sId).ciSubject._ptxCourseId) <$> ask
+evalSimpleTxsQuery (SELECTTxs _ (TxSubjectIdEq sId)) = do
+  db <- ask
+  return $ db ^..  traversed
+              . _SSTx
+              . filtered ((==sId).ciSubject._ptxCourseId)
 
 -- | Evaluator for query: find Txs in db with grade == g
-evalSimpleTxsQuery (SELECTTxs _ (TxGradeEq grade)) =
-  filter ((== Just grade).getGrade._ptxPayload) <$> ask
+evalSimpleTxsQuery (SELECTTxs _ (TxGradeEq grade)) = do
+  db <- ask
+  return $ db ^..  traversed
+              . _SSTx
+              . filtered ((== Just grade).getGrade._ptxPayload)
   where getGrade (EducatorTx (GradeCourse g)) = Just g
         getGrade _ = Nothing
 
 -- | Evaluator for query: find Txs in db with grade >= g
 evalSimpleTxsQuery (SELECTTxs _ (_ :>= grade)) = do
-  filter ((>= Just grade).getGrade._ptxPayload) <$> ask
+  db <- ask
+  return $ db ^..  traversed
+              . _SSTx
+              . filtered ((>= Just grade).getGrade._ptxPayload)
   where getGrade (EducatorTx (GradeCourse g)) = Just g
         getGrade _ = Nothing
 
@@ -71,10 +92,14 @@ evalSimpleTxsQuery (SELECTTxs _ (a :|| b)) =
    union <$> runQuery (SELECTTxs WHERE a) <*> runQuery (SELECTTxs WHERE b)
 
 -- | Evaluator for query: find all txs with subjectId which is descendant of sId
-evalSimpleTxsQuery (SELECTTxs _ (TxSubjectIsDescendantOf sId)) =
-  filter (isDescendantOf sId . ciSubject ._ptxCourseId) <$> ask
+evalSimpleTxsQuery (SELECTTxs _ (TxSubjectIsDescendantOf sId)) = do
+  db <- ask
+  return $ db ^..  traversed
+              . _SSTx
+              . filtered (isDescendantOf sId . ciSubject ._ptxCourseId)
   where isDescendantOf x y = hasPathFromTo activityTypeGraphIndexed x y
 
 -- | Run query in SimpleTxDB
 runSimpleTxDBQuery :: RunQuery a b => [PrivateTx] -> a -> b
-runSimpleTxDBQuery db query = runReader (getSimpleTxDB . runQuery $ query) db
+runSimpleTxDBQuery db query = runReader (getSimpleTxDB . runQuery $ query) mkDb
+ where mkDb = fmap SSTx db
